addressed some katex related comments
