+++ packages/unified-latex-to-pretext/tests/expand-user-defined-macros.test.ts
@@ -1,9 +1,8 @@
 import { describe, it, expect } from "vitest";
-import Prettier from "prettier";
 import util from "util";
 import { getParser } from "@unified-latex/unified-latex-util-parse";
-import { report_macros, expand_user_macros } from "../libs/report-and-expand-macros";
 import { printRaw } from "@unified-latex/unified-latex-util-print-raw";
+import { ExpandUserDefinedMacros } from "@unified-latex/unified-latex-to-pretext/libs/expand-user-defined-macros";
 
 // Make console.log pretty-print by default
 const origLog = console.log;
@@ -14,40 +13,13 @@ console.log = (...args) => {
 describe("unified-latex-to-pretext:report-and-expand-macro", () => {
     let value: string;
 
-    it("can reported unsupported macros", () => {
-        value = String.raw`$\mathbb{R} \fakemacro{X}$`;
-
-        const parser = getParser();
-        const ast = parser.parse(value);
-
-        expect(report_macros(ast)).toEqual(["fakemacro"]);
-    });
-
-    it("can report no unsupported macros in mathmode", () => {
-        value = String.raw`$\mathbb{R} \frac{1}{2} \cup$`;
-
-        const parser = getParser();
-        const ast = parser.parse(value);
-
-        expect(report_macros(ast)).toEqual([]);
-    });
-
-    it("can report no unsupported macros not in mathmode", () => {
-        value = String.raw`\underline{text} \textbf{bold} \subsection{section}`;
-
-        const parser = getParser();
-        const ast = parser.parse(value);
-
-        expect(report_macros(ast)).toEqual([]);
-    });
-
     it("can expand newcommand", () => {
         value = String.raw`\newcommand{\foo}{\bar{#1}}`;
 
         const parser = getParser();
         const ast = parser.parse(value);
 
-        expand_user_macros(ast)
+        ExpandUserDefinedMacros(ast);
 
         expect(printRaw(ast)).toEqual("\\bar{#1}");
     });
@@ -58,7 +30,7 @@ describe("unified-latex-to-pretext:report-and-expand-macro", () => {
         const parser = getParser();
         const ast = parser.parse(value);
 
-        expand_user_macros(ast)
+        ExpandUserDefinedMacros(ast);
 
         expect(printRaw(ast)).toEqual("\\N");
     });
@@ -69,7 +41,7 @@ describe("unified-latex-to-pretext:report-and-expand-macro", () => {
         const parser = getParser();
         const ast = parser.parse(value);
 
-        expand_user_macros(ast)
+        ExpandUserDefinedMacros(ast);
 
         expect(printRaw(ast)).toEqual("\\vee \\foo");
     });
